Parse 'section' logs + info about queueing duration
Provided a function that parses the log of freestyle jobs to
retrieve info about 'sections'.

A section is a log in the form of:
[section:<name>] start type=<type> time=<timestamp>
[section:<name>] end time=<timestamp>

For instance, it looks like:
[section:git-clone] start type=scm time=1527237016
[section:git-clone] end time=1527237026

This allows to represent what's happening inside a build even
though it is just a script.

This also adds a representation of the time spent in the queue as provided
by the https://github.com/jenkinsci/metrics-plugin/ plugin.

Resolves: FARM-1623
Change-Id: Ib6d161bdc9cdf54edfebc2535909318e2044cbd1